+++ packages/gakumas-engine/engine/TurnManager.js
@@ -147,6 +147,7 @@ export default class TurnManager extends EngineComponent {
 
     state[S.cardUsesRemaining] = 1;
 
+    state[S.prevStance] = "none";
     this.engine.effectManager.triggerEffects(state, FULL_POWER_EFFECTS);
     this.engine.effectManager.triggerEffectsForPhase(
       state,
@@ -195,9 +196,6 @@ export default class TurnManager extends EngineComponent {
     state[S.turnCardsUsed] = 0;
     state[S.turnCardsUpgraded] = 0;
     state[S.enthusiasm] = 0;
-    if (state[S.stance] == "fullPower") {
-      this.engine.buffManager.resetStance(state);
-    }
 
     this.engine.effectManager.decrementTtl(state);
     this.engine.effectManager.decrementDelay(state);